agregada una lista de ItemDescuento para el trackeo de descuentos sobre items

+++ TP2-Promociones/src/com/uba/tecnicas/promo/domain/Caja.java
@@ -30,7 +30,7 @@ public class Caja {
 		return ventaActual.getTotal();
 	}
 
-	public double getCantidadProductos() {
+	public int getCantidadProductos() {
 		return ventaActual.getCantidadProductosVendidos();
 	}
